tests/wpt/tests/css/css-grid/grid-lanes/tentative/parsing/grid-lanes-fill-valid.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Masonry: parsing masonry-fill with valid values</title>
  <link rel="author" title="Celeste Pan" href="mailto:celestepan@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-grid-3">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  <script>
    test_valid_value('masonry-fill', 'normal');
    test_valid_value('masonry-fill', 'reverse');
  </script>
</body>
</html>
